Refactored lambdas/ directory and updated API gateway paths
